Refactored global meta tags into a file
